// src/app/config/permissions-enum.ts
export enum EPermission {
  CHANGE_TICKET_STATE_PERMISSION = 'CHANGE_TICKET_STATE_PERMISSION',
  CHANGE_TICKET_PRIORITY_PERMISSION = 'CHANGE_TICKET_PRIORITY_PERMISSION',
  CHANGE_ACTIVITY_STATE_PERMISSION = 'CHANGE_ACTIVITY_STATE_PERMISSION',
  CHANGE_ACTIVITY_PRIORITY_PERMISSION = 'CHANGE_ACTIVITY_PRIORITY_PERMISSION',
  ADD_TICKET_CHECKLIST_PERMISSION = 'ADD_TICKET_CHECKLIST_PERMISSION',
  CHANGE_TICKET_CHECKLIST_STATE_PERMISSION = 'CHANGE_TICKET_CHECKLIST_STATE_PERMISSION',
  ADD_ACTIVITY_CHECKLIST_PERMISSION = 'ADD_ACTIVITY_CHECKLIST_PERMISSION',
  CHANGE_ACTIVITY_CHECKLIST_STATE_PERMISSION = 'CHANGE_ACTIVITY_CHECKLIST_STATE_PERMISSION',
  ADD_SUB_ACTIVITY_PERMISSION = 'ADD_SUB_ACTIVITY_PERMISSION',
  ADD_ACTIVITY_PERMISSION = 'ADD_ACTIVITY_PERMISSION',
  ADD_TICKET_COMMENT_PERMISSION = 'ADD_TICKET_COMMENT_PERMISSION',
  ADD_TICKET_ASSIGN_USER_PERMISSION = 'ADD_TICKET_ASSIGN_USER_PERMISSION',
  ADD_ACTIVITY_COMMENT_PERMISSION = 'ADD_ACTIVITY_COMMENT_PERMISSION',
  ADD_ACTIVITY_ASSIGN_USER_PERMISSION = 'ADD_ACTIVITY_ASSIGN_USER_PERMISSION',
  UPDATE_TICKET_PERMISSION = 'UPDATE_TICKET_PERMISSION',
  UPDATE_ACTIVITY_PERMISSION = 'UPDATE_ACTIVITY_PERMISSION',
  UPDATE_TICKET_CHECKLIST_PERMISSION = 'UPDATE_TICKET_CHECKLIST_PERMISSION',
  UPDATE_ACTIVITY_CHECKLIST_PERMISSION = 'UPDATE_ACTIVITY_CHECKLIST_PERMISSION',
  CREATE_FOLDER_PERMISSION = 'CREATE_FOLDER_PERMISSION',
  CREATE_CONTAINER_PERMISSION = 'CREATE_CONTAINER_PERMISSION',
  UPDATE_FOLDER_INFO = 'UPDATE_FOLDER_INFO',
  UPLOAD_ATTACHMENT = 'UPLOAD_ATTACHMENT',
  CAN_SEE_GROUP_MEMBERS = 'CAN_SEE_GROUP_MEMBERS',
  CAN_VALIDATE_TICKET_PERMISSION = 'CAN_VALIDATE_TICKET_PERMISSION',
  VIEW_TICKETS_PERMISSION = 'VIEW_TICKETS_PERMISSION',
  VIEW_ACTIVITIES_PERMISSION = 'VIEW_ACTIVITIES_PERMISSION',
  CHANGE_ACTIVITIY_TRACKTIME_PERMISSION = 'CHANGE_ACTIVITIY_TRACKTIME_PERMISSION',
}